#1049: Move interfaces to own files
