+++ CourageScores/ClientApp/src/interfaces/IServerSideError.ts
@@ -1,0 +1,5 @@
+import {IServerSideException} from "./IServerSideException";
+
+export interface IServerSideError {
+    Exception?: IServerSideException;
+}
