+++ CourageScores/ClientApp/src/interfaces/IServerSideException.ts
@@ -1,0 +1,6 @@
+export interface IServerSideException {
+    Type: string;
+    StackTrace?: string[];
+    Message: string;
+}
+
